change deploy scripts

+++ scripts/stop-compose.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
-
-docker compose down
-docker rmi -f $(docker images -aq)
+if [ -d /usr/app/code-engine-server ]; then
+  cd /usr/app/code-engine-server
+  docker compose down
+  docker rmi -f $(docker images -aq)
+fi
